Refatoramento de codigo
Tudo funcionando, mas precisa verificar se realmente e' o que esta'
pedindo
