+++ atividade1/questao8.py
@@ -1,0 +1,13 @@
+from util import *
+
+def questao8_sobel(img):
+	'''
+	Filtro de deteccao de bordas Sobel
+	'''
+	pass
+
+def questao8_laplaciano(img):
+	'''
+	Filtro de deteccao de bordas laplaciano
+	'''
+	pass
