+++ atividade1/questao9.py
@@ -1,0 +1,28 @@
+from util import *
+
+def questao9_a(img):
+	'''
+	aplicar o filtro
+	  0  -1	   0
+		+	+
+	 -1	  5   -1
+		+	+
+	  0	  -1   0
+
+	'''
+	applyFilter(img, [[],[],[]])
+
+	pass
+
+def questao9_b(img):
+	'''
+	aplicar o filtro
+	  0   0	   0
+		+	+
+	  0	  1    0
+		+	+
+	  0	   0   -1
+	'''
+
+	applyFilter(img, [[],[],[]])
+	pass
